requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Werkzeug==2.3.7
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
deepface==0.0.79
tensorflow==2.13.0
tf-keras==2.13.1